+++ CHANGELOG.md
@@ -13,6 +13,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `New-DscCompositeResourceWikiPage` - Returns the markdown content for a
     wiki page for a DSC composite resource.
 
+### Changed
+
+- `New-DscResourceWikiPage`
+  - Added support for creating wiki pages for composite resources.
+
 ## [0.9.0] - 2021-07-08
 
 ### Added